+++ autologin_platform_interface/test/autologin_platform_interface_test.dart
@@ -40,6 +40,12 @@ class AutologinMock extends AutologinPlatform {
     loginToken = token;
     return true;
   }
+
+  @override
+  Future<bool> deleteLoginToken() async {
+    loginToken = null;
+    return true;
+  }
 }
 
 void main() {