+++ autologin_web/lib/autologin_web.dart
@@ -57,7 +57,8 @@ class AutologinWeb extends AutologinPlatform {
   }
 
   @override
-  Future<bool> saveLoginToken(String token) async {
-    return false;
-  }
+  Future<bool> saveLoginToken(String token) async => false;
+
+  @override
+  Future<bool> deleteLoginToken() async => false;
 }