+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+* Add API to delete saved autologin tokens (supported on Android, iOS and macOS)
+* Add Swift Package Manager support
+* Extend example app for publishing on Google Play
+
 ## 1.0.0
 
 * First official stable version