+++ autologin_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+* Add API to delete saved autologin tokens
+
 ## 1.0.0
 
 * First official stable version